requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
fake-useragent==0.1.11
idna==2.10
imageio==2.9.0
jellyfish==0.8.2
kiwisolver==1.2.0
lxml==4.6.2
matplotlib==3.3.1
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
parse==1.16.0
Pillow==8.1.0
proxyscrape==0.3.0
psutil==5.7.2
pyee==7.0.2
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.3
pytesseract==0.3.7
python-dateutil==2.8.1
PyWavelets==1.1.1
ratelimit==2.2.1
requests==2.24.0
requests-html==0.10.0
scikit-image==0.17.2
scipy==1.5.3
six==1.15.0
soupsieve==2.0.1
tifffile==2020.10.1
tqdm==4.48.2
urllib3==1.25.10
w3lib==1.22.0
websockets==8.1
nest-asyncio==1.5.5
